pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "biblioteca"
version = "0.0.1"
authors = [
  { name="Matheus Costa Monteiro", email="matheuscostamonteiro.mc@ufu.br" },
  { name="Thiago Pacheco Rocha", email="thiago.rocha@ufu.br" },
]
description = "Sistema de biblioteca para projeto da disciplina sistemas distribuídos."
readme = "README.md"
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.11"
dependencies = [
   "grpcio",
]

[project.optional-dependencies]
dev = []

[project.scripts]
cad-server="biblioteca.cad.cad_server:run"
cad-client="biblioteca.cad.cad_client:run"
bib-server="biblioteca.bib_server:run"
bib-client="biblioteca.bib_client:run"

[project.urls]
Homepage = "https://github.com/costadev00/sistemas-distribuidos-projeto"
Issues = "https://github.com/costadev00/sistemas-distribuidos-projeto/issues"
